Added log in and register account functionality before app is able to be used. Refactored code.
