+++ lib/screens/login.dart
@@ -1,6 +1,6 @@
+import 'package:adoption_hero/screens/navigator_scaffold.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:adoption_hero/main.dart';
 import 'package:adoption_hero/screens/register.dart';
 
 class LoginWidget extends StatefulWidget {
@@ -14,84 +14,108 @@ class _LoginWidgetState extends State<LoginWidget> {
   final TextEditingController _emailField = TextEditingController();
   final TextEditingController _passwordField = TextEditingController();
 
+
+  @override
+  void dispose(){
+    _emailField.dispose();
+    _passwordField.dispose();
+
+    super.dispose();
+  }
+
   @override
   Widget build(BuildContext context) {
-    return Padding(
-        padding: const EdgeInsets.all(10),
-        child: ListView(
-          children: <Widget>[
-            Container(
-                alignment: Alignment.center,
+    return Scaffold(backgroundColor: Colors.green[50],
+      appBar: AppBar(
+        title: const Icon(Icons.pets_rounded),
+        backgroundColor: Colors.green[100],
+      ),
+      body: Padding(
+          padding: const EdgeInsets.all(10),
+          child: ListView(
+            children: <Widget>[
+              Container(
+                  alignment: Alignment.center,
+                  padding: const EdgeInsets.all(10),
+                  child: const Text(
+                    'Adoption Hero',
+                    style: TextStyle(
+                        color: Colors.green,
+                        fontWeight: FontWeight.w500,
+                        fontSize: 30),
+                  )),
+              Container(
+                  alignment: Alignment.center,
+                  padding: const EdgeInsets.all(10),
+                  child: const Text(
+                    'Sign in',
+                    style: TextStyle(fontSize: 20),
+                  )),
+              Container(
                 padding: const EdgeInsets.all(10),
-                child: const Text(
-                  'Adoption Hero',
-                  style: TextStyle(
-                      color: Colors.green,
-                      fontWeight: FontWeight.w500,
-                      fontSize: 30),
-                )),
-            Container(
-                alignment: Alignment.center,
-                padding: const EdgeInsets.all(10),
-                child: const Text(
-                  'Sign in',
-                  style: TextStyle(fontSize: 20),
-                )),
-            Container(
-              padding: const EdgeInsets.all(10),
-              child: TextFormField(
-                controller: _emailField,
-                decoration: const InputDecoration(
-                  border: OutlineInputBorder(),
-                  labelText: 'Email',
+                child: TextFormField(
+                  controller: _emailField,
+                  decoration: const InputDecoration(
+                    border: OutlineInputBorder(),
+                    labelText: 'Email',
+                  ),
                 ),
               ),
-            ),
-            Container(
-              padding: const EdgeInsets.fromLTRB(10, 10, 10, 0),
-              child: TextFormField(
-                obscureText: true,
-                controller: _passwordField,
-                decoration: const InputDecoration(
-                  border: OutlineInputBorder(),
-                  labelText: 'Password',
+              Container(
+                padding: const EdgeInsets.fromLTRB(10, 10, 10, 0),
+                child: TextFormField(
+                  obscureText: true,
+                  controller: _passwordField,
+                  decoration: const InputDecoration(
+                    border: OutlineInputBorder(),
+                    labelText: 'Password',
+                  ),
                 ),
               ),
-            ),
-            TextButton(
-              onPressed: () {
-                // forgot password screen
-              },
-              child: const Text('Forgot Password',),
-            ),
-            Container(
-                height: 50,
-                padding: const EdgeInsets.fromLTRB(10, 0, 10, 0),
-                child: ElevatedButton(
-                  child: const Text('Login'),
-                  onPressed: () {
-                  },
-                )
-            ),
-            Container(
-                height: 50,
-                padding: const EdgeInsets.fromLTRB(10, 0, 10, 0),
-                child: ElevatedButton(
-                  child: const Text('Register'),
-                  onPressed: () async {
-                    await register(_emailField.text, _passwordField.text);
-                  },
-                )
-            ),
-          ],
-        ));
+              TextButton(
+                onPressed: () {
+                  // forgot password screen
+                },
+                child: const Text('Forgot Password',),
+              ),
+              Container(
+                  height: 50,
+                  padding: const EdgeInsets.fromLTRB(10, 0, 10, 0),
+                  child: ElevatedButton(
+                    child: const Text('Login'),
+                    onPressed: () async {
+                      bool shouldNavigate = await signIn();
+                      if(shouldNavigate){
+                        Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (context) => const NavigatorScaffold()));
+                      }
+                    },
+                  )
+              ),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: <Widget>[
+                  const Text('Don\'t have an account?'),
+                  TextButton(
+                    child: const Text(
+                      'Register',
+                      style: TextStyle(fontSize: 20),
+                    ),
+                    onPressed: () {
+                      Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (context) => const RegisterWidget()));
+                    },
+                  )
+                ],
+              ),
+            ],
+          )),
+    );
   }
-}
 
-Future<bool> register(String email, String password) async {
-try{
+  Future<bool> signIn() async {
+  try{
     await FirebaseAuth.instance
-    .createUserWithEmailAndPassword(email: email, password: password);
+    .signInWithEmailAndPassword(email: _emailField.text.trim(), password: _passwordField.text.trim());
+    print("logged in");
     return true;
   } on FirebaseAuthException catch (e) {
     if(e.code == 'weak-password'){
@@ -102,6 +126,10 @@ try{
     return false;
   } catch (e) {
     print(e.toString());
-    return false;
   }
+  return false;
 }
+}
+
+
+
